Make util module lighter by moving some functions
