+++ labonneboite/common/util.py
@@ -2,14 +2,11 @@
 
 from urlparse import urlparse
 import logging
-import re
 import unicodedata
 import urllib
 
-from flask import request, session
-from flask_login import current_user
+from flask import request
 
-from labonneboite.conf import settings
 from labonneboite.conf.common.contact_mode import CONTACT_MODE_DEFAULT
 from labonneboite.common.load_data import load_contact_modes
 
@@ -50,56 +47,6 @@ def get_user_ip():
     # otherwise, fallback to remote_addr.
     # http://werkzeug.pocoo.org/docs/0.10/wrappers/#werkzeug.wrappers.BaseRequest.access_route
     return request.access_route[0] if request.access_route else request.remote_addr
-
-
-def user_is_pro():
-    """
-    The website has a special version called "Version PRO"
-    which is only visible to "PRO users"
-
-    PRO users are
-    - Conseillers PE identified by their specific IP
-    - Institutionals identified by their specific PEAM emails
-
-    The "Version PRO" is the same as the regular "Version publique",
-    plus some exclusive indicators, filters and data:
-    - Flags : Junior, Senior, Handicap...
-    - Office data : statistics about recruitments...
-    """
-
-    # Check user IP (no need to be authenticated)
-    user_ip = get_user_ip()
-    if user_ip in settings.VERSION_PRO_ALLOWED_IPS:
-        return True
-
-    # Check user e-mail by plain_value, suffix or regex (@see local_settings.py)
-    if current_user.is_authenticated:
-        current_user_email = current_user.email.lower()
-
-        return (current_user_email in settings.VERSION_PRO_ALLOWED_EMAILS
-            or any(current_user_email.endswith(suffix) for suffix in settings.VERSION_PRO_ALLOWED_EMAIL_SUFFIXES)
-            or any(re.match(
-                regexp, current_user_email) is not None for regexp in settings.VERSION_PRO_ALLOWED_EMAIL_REGEXPS))
-
-    return False
-
-def pro_version_enabled():
-    if not user_is_pro() and 'pro_version' in session:
-        session.pop('pro_version')
-    return session.get('pro_version', False)
-
-def get_doorbell_tags(tag):
-    if tag not in ['faq', 'help', 'press', 'results']:
-        raise Exception('unknown page')
-    if user_is_pro():
-        return ['conseiller', tag]
-    return ['de', tag]
-
-
-def get_hotjar_tag():
-    if user_is_pro():
-        return "conseiller"
-    return "de"
 
 
 def sanitize_string(s):
@@ -156,4 +103,3 @@ def get_contact_mode_for_rome_and_naf(rome, naf):
         return naf_prefix_to_rome_to_contact_mode[naf_prefix].values()[0]
     except (KeyError, IndexError):
         return CONTACT_MODE_DEFAULT
-
