+++ labonneboite/tests/web/integration/test_root.py
@@ -2,7 +2,7 @@
 import mock
 
 from labonneboite.tests.test_base import DatabaseTest
-from labonneboite.common import util
+from labonneboite.common import pro
 from labonneboite.common.models import User
 
 class RootTest(DatabaseTest):
@@ -10,8 +10,8 @@ class RootTest(DatabaseTest):
     def login_as_pro(self):
         user_pro = User.create(email=u'x@pole-emploi.fr', gender=u'male', first_name=u'John', last_name=u'Doe')
         self.login(user_pro)
-        self.assertTrue(util.user_is_pro())
-        self.assertFalse(util.pro_version_enabled())
+        self.assertTrue(pro.user_is_pro())
+        self.assertFalse(pro.pro_version_enabled())
 
     def test_no_kit_if_public_user(self):
         rv = self.app.get(self.url_for('root.kit'))
